# ARROW-7177: [Java] Provide a utility to improve the performance of vector loading/unloading