--- a/java/vector/src/main/java/org/apache/arrow/vector/TypeLayout.java
+++ b/java/vector/src/main/java/org/apache/arrow/vector/TypeLayout.java
@@ -231,8 +231,15 @@ public class TypeLayout {
   public static int getTypeBufferCount(final ArrowType arrowType) {
     return  arrowType.accept(new ArrowTypeVisitor<Integer>() {
 
+      /**
+       * All fixed width vectors have a common number of buffers 2: one validity buffer, plus a data buffer.
+       */
       static final int FIXED_WIDTH_BUFFER_COUNT = 2;
 
+      /**
+       * All variable width vectors have a common number of buffers 3: a validity buffer,
+       * an offset buffer, and a data buffer.
+       */
       static final int VARIABLE_WIDTH_BUFFER_COUNT = 3;
 
       @Override
@@ -247,6 +254,7 @@ public class TypeLayout {
             // TODO: validate this
             return 3;
           case Sparse:
+            // type buffer
             return 1;
           default:
             throw new UnsupportedOperationException("Unsupported Union Mode: " + type.getMode());
@@ -255,6 +263,7 @@ public class TypeLayout {
 
       @Override
       public Integer visit(Struct type) {
+        // validity buffer
         return 1;
       }
 
@@ -265,16 +274,19 @@ public class TypeLayout {
 
       @Override
       public Integer visit(org.apache.arrow.vector.types.pojo.ArrowType.List type) {
+        // validity buffer + offset buffer
         return 2;
       }
 
       @Override
       public Integer visit(FixedSizeList type) {
+        // validity buffer
         return 1;
       }
 
       @Override
       public Integer visit(Map type) {
+        // validity buffer + offset buffer
         return 2;
       }