Vector loading/unloading transforms a set of vectors to and from a set of buffers with meta data. It is heavily used in flight/IPC.

In the loading/unloading operations, only the number of type buffers are really needed. However, the current code logic gets a copy of the type buffers, which is not necessary.

In this issue, we provide a utility to get the number of type buffers, given an arrow type. It improves the performance by

1. avoiding creating objects unnecessarily.
2. avoiding list copying for vector unloading (which calls TypeLayout#getBufferTypes).