@lidavidm 2019-11-15
If I understand correctly, the original code was not copying any data, but copying (potentially) many objects describing the layout of the buffers?

Are there benchmark results for this change? (I haven't quite kept up with the new benchmark infrastructure...) I'd be curious if this makes a noticeable difference compared to the cost of actually reading/writing data in IPC or Flight.