@liyafan82 2019-11-18
@lidavidm @siddharthteotia Thanks a lot for your comments.

I have provided a simple benchmark to evaluate the performance for vector unloading, because the performance improvement is greater for vector unloading. The results show only trivial performance improvement:

Before:
VectorUnloaderBenchmark.unloadBenchmark  avgt    5  731.253 ± 32.402  us/op

After:
VectorUnloaderBenchmark.unloadBenchmark  avgt    5  719.044 ± 12.481  us/op

Further investigations show the reason is that most time (more than 95%) is spent in creating the ArrowRecordBatch object. I think that problem deserves a separate issue. What do you think?