@liyafan82 2019-11-18
> If I understand correctly, the original code was not copying any data, but copying (potentially) many objects describing the layout of the buffers?

You are right. In the old implementation, it creates a new list, and copy each element of the buffer object to the new list. So only shallow copy is performed.