@lidavidm 2019-11-21
For such a small performance improvement, I'm not sure this is worth it - in fact, the two distributions overlap! If indeed the vast majority of the time is in ArrowRecordBatch's constructor, it'd probably be better to focus efforts there.