@liyafan82 2019-11-25
> For such a small performance improvement, I'm not sure this is worth it - in fact, the two distributions overlap! If indeed the vast majority of the time is in ArrowRecordBatch's constructor, it'd probably be better to focus efforts there.

@lidavidm Thanks for your valuable feedback, and sorry for my late response.

After investigation, we have found the reason for the long running time of creating a ArrowRecordBatch.

Briefly speaking, we have turned on the BaseAllocator#DEBUG flag by accident, so HistoricalLog#recordEvent is invoked every time the Reference#retain method is called. 

HistoricalLog#recordEvent is an expensive operation, and only used for debugging purpose. So we have disabled this operation in our benchmark. 

The new performance evaluation shows a 30+% performance improvement:

after:
VectorUnloaderBenchmark.unloadBenchmark  avgt    5  1.668 ± 0.152  us/op

before:
VectorUnloaderBenchmark.unloadBenchmark  avgt    5  2.513 ± 0.462  us/op