@liyafan82 2019-11-26
> Ah, thank you for investigating further. I think that makes the improvement more reasonable, but I'd be curious what effect it has on the overall runtime of reading a stream (via Flight or otherwise).
> 
> Additionally, given that it's common in Flight to unload and then immediately load a batch, I wonder if the benchmark of the two operations together shows noticeable improvement.

Sure. I am also interested in the result. However, maybe the improvement is not that great, as this change only improves a small part of the overall process. 

I will conduct more experiments later, and post the results here.