@liyafan82 2019-11-29
> Ah, thank you for investigating further. I think that makes the improvement more reasonable, but I'd be curious what effect it has on the overall runtime of reading a stream (via Flight or otherwise).
> 
> Additionally, given that it's common in Flight to unload and then immediately load a batch, I wonder if the benchmark of the two operations together shows noticeable improvement.

@lidavidm Sorry for my late response. 

I have added a benchmark for vector loading. The performance improvement is about 8%, not as great as the improvement for vector unloading:

after
VectorLoaderBenchmark.loadBenchmark  avgt    5  4.418 ± 0.354  us/op

before
VectorLoaderBenchmark.loadBenchmark  avgt    5  4.800 ± 0.452  us/op

This is because, the origin implementation of vector loading does not involve the overhead of creating/copying type list (it was based on TypeLayout#getTypeLayout).

So we have the following overall results:


Benchmark | Before (us) | After (us) | Improvement
-- | -- | -- | --
Unloading | 2.513 | 1.668 | 33.6%
Loading | 4.800 | 4.418 | 8.0%
Unloading + loading | 7.313 | 6.106 | 16.5%