@liyafan82 2019-11-29
I also evaluated the performance by TestPerf tool in flight module. The evaluation method is: run one round of warmup, and take the best of consecutive two runs. 

Before:
Average throughput: 1253.360341 MiB/s, standard deviation: 120.045803 MiB/s

After:
Average throughput: 1272.005214 MiB/s, standard deviation: 164.509097 MiB/s

However, I don't think the result is as reliable as the result obtained from JMH, as results of different runs may vary significantly.