@liyafan82 2019-12-06
> @liyafan82 sorry for the delay. The improvement looks reasonable, but I'm not convinced that it makes a difference outside the microbenchmarks. (Maybe plain IPC is a better way to test this, instead of using Flight which has to hit the network?)

@lidavidm Thanks for your valuable feedback. I agree with you that the performance results may be different in real scenarios, as our microbenchmarks only evaluate the performance of serialization/deserialization for vectors, which is only part of the whole process.

As you have indicated, the results of Flight may distort the results, due to the network cost. There are other ways of IPC (e.g. shared memory, pipe, etc.). However, there also some extra cost (e.g. memory copy, system calls, etc.). Do you have any target scenario to support in mind?