@lidavidm 2019-12-06
Perhaps just repeatedly loading/writing the same data from/to an in-memory buffer would be enough - that way we include IPC protocol cost but not a file/network/domain socket.

(To be clear: I'm OK with the PR as-is, though I worry a little about keeping the visitor in sync with future changes.)