@liyafan82 2019-12-09
> Perhaps just repeatedly loading/writing the same data from/to an in-memory buffer would be enough - that way we include IPC protocol cost but not a file/network/domain socket.
> 
> (To be clear: I'm OK with the PR as-is, though I worry a little about keeping the visitor in sync with future changes - though there is the unit test.)

Thanks for your additional feedback. You concern makes sense to me, and we should be mindful of the related future changes.