@lidavidm 2019-12-09
Perhaps someone more familiar with Java internals can take a look too?