@siddharthteotia 2019-12-12
LGTM -- although I do feel this change may have a little impact on the overall pipeline of reading/writing via Flight as this may not necessarily be the bottleneck. So I feel it is unlikely to move the needle significantly in such cases.

Having said that, I am okay with merging the PR. @liyafan82 , please update the PR description since it is misleading and indicating that current code was doing some sort of memcpy whereas a list is created with objects  (buffers) whereas we only need the number of buffers